Tue, Feb 20, 2024, 11:33 AM +05:30

+++ main.cpp
@@ -10,7 +10,7 @@ int main()
     cin >> a;
     cin >> b;
 
-    cout << fixed << (float)a + b << endl;
+    cout << fixed << (float)a + (float)b << endl;
 
     average = ((float)a + b) / 2;
